backend/requirements.txt
langchain==0.1.11
openai
tiktoken
faiss-cpu
msal
pylint
autopep8
black
azure-functions
langchain-openai
langchain-community
pytest
flask
pandas
psycopg[binary]
pgvector
openpyxl
azure-core
python-docx
pypdf